feat: change keccak256 opcode to write output as bytes (#396)
For most purposes, we just convert the digest u16 limbs into bytes
anyways. I realized it is possible to do this for free with the way the
Keccak chip is currently implemented.

Safety Note: the output bytes are not range checked. (This was also a
problem with the u16 to u8 decomposition before this change so it is a
separate concern.)