[fix] RV32M DIV: division by zero not fully constrained (#33)
**Finding Link:**
https://cantina.xyz/code/c486d600-bed0-4fc6-aed1-de759fd29fa2/findings/53

## Description of Fix
In the `DivRemCoreChip`, there is a `zero_divisor` flag to indicate when
`c = 0`, ensuring that `q = -1` is enforced if the flag is set. However,
this flag was not constrained. So if `c = 0` but a malicious prover sets
`zero_divisor = 0`, there would be no constraint ensuring the correct
value for `q`.

This pull request addresses that issue by adding a new constraint:

> If `c = 0` then `zero_divisor = 1`  
> *(or equivalently, if `zero_divisor = 0` then `c != 0`)*

To check whether `c = 0`, we use `c_sum`, which is the sum of `c`’s
limbs. We also add a new column, `c_sum_inv`, which verifies `c_sum` is
nonzero similar to how `IsZeroSubAir` works.

Because `r_zero` has the same issue, this PR applies an equivalent fix
for `r_zero` as well.

Resolves INT-3408

---------

Co-authored-by: Jonathan Wang <31040440+jonathanpwang@users.noreply.github.com>