[fix] Add the not-overflowing assertion to the range tuple checker bus ctor (#49)
**Finding Link:**
https://cantina.xyz/code/c486d600-bed0-4fc6-aed1-de759fd29fa2/findings/90

## Description of Fix

Now, on the range tuple checker bus creation, we check that the product
of the sizes (aka height of the preprocessed trace) will not overflow
`u32`, and there also is a test that checks the assertion trigger.
While we could generally check that the product doesn't exceed something
more meaningful and not just 2^32, this is out of scope of the finding
since it would probably depend on unrelated things, such as security
parameters and backend theorems, and therefore would be applied
somewhere else.