fix: Poseidon2 verify_batch `start_top_level` is not constrained only at start (#1403)
**Finding Link:**
https://cantina.xyz/code/c486d600-bed0-4fc6-aed1-de759fd29fa2/findings/167

### Description of Fix

While the variable was called `start_top_level`, we only constrained
that the top level has `start_top_level = 1` but never that it must be
equal to `0` on non-top levels. We change the existing constraint so
that it enforces `start_top_level = 1` iff its the start of a top level
(meaning previous row is `end` and current row is `incorporate_row`).

Closes INT-3567