fix: Fix bug in setup row of modular equality chip (`ModularIsEqualCoreChip`) (#1409)
This PR fixes a bug in the algebra extension that was found during the
internal security review.

**Description:** The setup row of the `ModularIsEqualCoreChip` is
unsound. It does not correctly constrain that the modulus is correct.
The problem is that the casework on `prefix_sum` that allows
constraining `b[i] = N[i]` for the correct indices (when `i >
b_diff_idx`) fails on the setup row. The intended behavior on the setup
row is that `c_lt_mark = 2` and `lt_marker` has exactly one 2 (at
`c_diff_idx`) and the remaining entries are 0s.

There are three ways in which the setup row is underconstrained:
1. `c_lt_mark` is not constrained to be 2 
2. Even if `c_lt_mark = 2`, `sum_i lt_marker[i]` is not constrained to
be 2 (which would ensure that `lt_marker` has exactly one 2 and that the
remaining entries are 0s).
3. Even if the above two constraints hold, `b[i] = N[i]` is only
constrained for `i > c_diff_idx` while it should be constrained for all
`i`.

See the failing tests added in the first commit of this PR for a proof
of concept.

Closes INT-3599