fix: complete `verify_segments` functionality (#1438)
**Finding Links:**
-
https://cantina.xyz/code/c486d600-bed0-4fc6-aed1-de759fd29fa2/findings/39
-
https://cantina.xyz/code/c486d600-bed0-4fc6-aed1-de759fd29fa2/findings/56

### Description of Fix

- Move the logic to compute exe commit into `openvm_circuit` crate.
- Finish `verify_segments` function but where exe commitment and final
memory root are returned.
- Add `UserPublicValuesProof::verify` function for verifying user public
values with respect to the final memory root. Remove the left/right
indicators from the proof because they **must** be determined from the
address space + memory dimensions
- Add `MemoryDimensions` to `AppVerifyingKey` for the above reason
- The combination of `verify_segments` and user public values verify is
equivalent in functionality to the full recursive aggregation tree down
to the root proof verification, which is now what the SDK
`verify_app_proof` does.
- To get some stuff to work with generics, `verify_segments` and
`verify_single` are now standalone functions not part of
`VirtualMachine`.

Closes INT-3286

---------

Co-authored-by: Jonathan Wang <31040440+jonathanpwang@users.noreply.github.com>