chore(refactor): separate `openvm-continuations` from SDK (#1447)
A refactor to split the security critical recursion programs necessary
for the continuations framework outside of the SDK. The recursion
programs are now in a separate `openvm-continuations` crate.

I wanted to keep keygen together with the recursion programs, but that
was too difficult to separate because the proving key types are very
tied to the SDK, and also the requirement of a dummy proof ties keygen
to some prover implementation. This is not ideal and could be reworked
with more refactoring, which is out of scope for now.

No major logic was changed. Most files changed are import changes.
- The `static_verifier` module was split between `sdk/keygen` and
`continuations/verifier`, with some minor changes to remove dependence
on proving key in the verifier (which is weird for a verifier anyways).
- `ContinuationVmProof` is moved into `openvm-circuit` crate to remove a
dependency issue and also because it is core to the framework.