fix: Add missing shape validation into recursion program (#1464)
The shape of `opening.values.preprocessed` is not fully validated. When
preparing the rounds for preprocessed traces
[here](https://github.com/openvm-org/openvm/blob/18096f4194b76b743463c0fb39955f24d010e9bf/extensions/native/recursion/src/stark/mod.rs#L405)
could access an out-of-bound index.


[Here](https://github.com/openvm-org/openvm/blob/18096f4194b76b743463c0fb39955f24d010e9bf/extensions/native/recursion/src/stark/mod.rs#L413)
lengths of `prep.local` and `prep.next` should equal to the width of the
preprocessed trace.

The shape of `opening.values.after_challenge` is also not fully
validated.
[Here](https://github.com/openvm-org/openvm/blob/18096f4194b76b743463c0fb39955f24d010e9bf/extensions/native/recursion/src/stark/mod.rs#L554)
the width of trace is not validated.

closes INT-3644