fix: completeness of ModBuilder trace generation (#1473)
Previously the trace generation for ModBuilder was not calculating the
number of limbs necessary in a way consistent with the constraints. This
was because there were some underestimates when the prime is much
smaller than $`2^\textrm{num\_limbs * limb\_bits}`$. We fix it by
updating all calculations to use the theoretical bound that an input or
saved var have big integer value in $`[0, 2^\text{num\_limbs *
limb\_bits})`$ rather than $[0, p)$.

This causes no soundness bugs, but it does affect completeness of
tracegen.

closes INT-3641