chore: Range check carries to fixed value (#1475)
`save_if_overflow` previously saved the expression if its carries could
be larger than `range_checker_bits`. However, this means
`range_checker_bits` must be small enough to not cause overflow. Really,
the condition to do a save should be if the carries could be larger than
`min(range_checker_bits, max_carry_bits)` for a `max_carry_bits` value
that ensures no overflow.

I added this `max_carry_bits` value to be `31 - limb_bits - 2`. This
ensures that the carries are in $\left(-\frac{p}{2}, \frac{p}{2}\right)$
where $p$ is the BabyBear modulus which ensures the carry constraints do
not overflow. See the code comment in `check_carry_to_zero.rs` for a
justification of this bound.

Closes INT-3533