fix: Add `bls12_381` feature to clippy command in lint CI (#1480)
The lint workflow currently runs clippy on guest programs used as
integration tests. It special cases the ecc and pairing integration test
guest programs and runs clippy with certain features turned on.

I noticed that when clippy lints the pairing guest programs with the
`bn254` feature turned on, but not with the `bls12_381` feature turned
on. Due to conditional compilation used in the guest code, this results
in some of the code not being linted.

This PR duplicates the clippy command with the `bls12_381` feature
turned on. The reason that we don't enable both `bn254` and `bls12_381`
in a single clippy run is that some tests use the `cfg_match` macro to
select between curves, and, as a result, the test for only one curve is
compiled.

Also fixes the linter errors that this change reveals.

Closes INT-3655