docs: update specs on RISC-V execution environment (#1489)
Updates specs on the RISC-V execution environment for user IO.

Update internal docs to explain soundness of hint buffer. This relies on
an invariant of the memory bus which is stated in
https://github.com/openvm-org/openvm/pull/1490.

Also found that our `read_vec_with_len` implementation uses a bit of
unsafe Rust that does not quite match the documented safety assumptions
(though in practice currently does not have any problems). This is
around a hidden assumption that the allocator always allocates on a
4-byte boundary.