feat(recursion): add clarifying OOB check (#1503)
In recursion program, add a check that the number of after challenge
commitments equals the number of phases as recorded in the vkey. This is
just for clarity and not a soundness fix: the control flow is already
determined by `num_phases` from the vkey, and even if an out of bounds
memory access were made, it is equivalent to if an array of the correct
length 1 were provided, but where the value was incorrect -- this would
still be caught by the verification algorithm.

Add some more comments in recursion program.

Updates stark-backend commit to sync with
https://github.com/openvm-org/stark-backend/pull/57