feat: `GenericSdk<E: StarkFriEngine<BabyBearPoseidon2Config>>` (#1519)
This change introduces a generic type for Sdk (with a typedef for `Sdk`
that uses `BabyBearPoseidon2Engine`). This is to allow Sdk usage with
different engines that still use the same `StarkGenericConfig =
BabyBearPoseidon2Config`.

This will require changes in other repos, but it is inevitable since
`let sdk = Sdk;` is now an invalid assignment.

---------

Co-authored-by: Jonathan Wang <31040440+jonathanpwang@users.noreply.github.com>