chore(ecc-guest): placeholder trait and function implementations for patch compilation (#1720)
The `alloy-primitives` and `alloy-consensus` crates turn on ECDSA
signing and verification algorithms together when `"k256"` feature is
enabled. In order for our `k256` crate to be used as a patch, I had to
provide functions to match.

Because we use a custom `VerifyingKey` struct (unavoidable since the
underlying types are custom to `openvm`), the `SigningKey` struct must
also be custom (the required methods fundamentally will not work if we
try to use the `ecdsa` crate's `SigningKey` but our own `VerifyingKey`).
I have left the implementations `todo!` because we do not yet support
signing as a functionality -- I could try to do a byte conversion to the
`ecdsa::SigningKey` to passthrough the implementation, but I wanted to
reduce the security surface area for now.

Fixed a regression in ecrecover benchmark because the crates.io version
of `revm` always set `sha3-keccak` feature on for `revm-precompile`
which was turning off `native-keccak` linkage.