fix: use seed for deterministic re-execution (#1741)
The `NonQrHintSubEx` phantom sub-executor now uses a fixed seed so that
execution is deterministic between different instances of the VM. This
is not strictly necessary since the constructed non-QR could be passed
between runs for reproducibility, but it seems a better property that
re-runs of the VM have the same behavior (e.g., for segmentation).

The only additional consideration is that now you can potentially add
moduli to the VM config that take a little longer for the VM to
initialize - but making it truly take a long time seems to involve
breaking generalized Riemann hypothesis.

Lastly, the phantom sub-executors `DecompressHintSubEx` and
`NonQrHintSubEx` in Weierstrass extension were already removed from the
spec since we now use the phantom sub-executors in the modular
arithmetic extension, but these variants were not deleted from the code.
So they are deleted in this PR. @Avaneesh-axiom do you know if this was
from a bad merge previously?