fix(k256): ecdsa verify needs custom hook (#1744)
For some reason, unlike other libraries, the k256 crate has a special
requirement that the signature is normalized ahead of time:
https://github.com/RustCrypto/elliptic-curves/blob/5ac8f5d77f11399ff48d87b0554935f6eddda342/k256/src/ecdsa.rs#L203
To handle this, we introduce a new trait `VerifyCustomHook` that is
called both in `recover_prehashed` and `verify_prehash`.

- [x] Provide a proof that valid pubkey recovery implies the ECDSA
signature verification algorithm will automatically pass, so there is no
need for `recover_prehashed` to call `verify_prehash` again.